Create project details container and move project settings form logic into details container
