+++ client/src/containers/ProjectDetails/logic/reducer.ts
