+++ client/src/containers/ProjectDetails/selectIcon.tsx
