+++ client/src/containers/ProjectDetails/styles.module.scss
@@ -1,0 +1,121 @@
+.header_inner__container {
+    padding: 10px 0px 20px ;
+    position: relative;
+}
+
+.header__options {
+    position: absolute;
+    right: 30px;
+    top: -33px;
+}
+
+.header_inner__title {
+    font-size: 24px;
+    margin: 0;
+}
+.body_inner__container {
+    position: relative;
+}
+
+.form__container {
+    width: 580px;
+    margin: 0 auto;
+    margin-bottom: 25px;
+}
+
+.fix_start_after_input {
+    &::after {
+        content: ''!important;
+    }
+}
+
+.form__input {
+    @extend .fix_start_after_input;
+    width: 250px;
+}
+
+.form__input_key{
+    @extend .fix_start_after_input;
+    width: 125px;
+}
+
+.form__avatar {
+    display: flex;
+    align-items: center;
+    cursor: pointer;
+    border: none;
+    background-color: transparent;
+    color: rgb(0, 101, 255);
+
+    &:hover {
+        text-decoration: underline;
+    }
+}
+
+.avatar__img {
+    height: 40px;
+    min-width: 40px;
+    margin-right: 10px;
+    border-radius: 3px;
+}
+
+.form__area {
+    @extend .fix_start_after_input;
+}
+
+.form__input_container {
+    display: flex;
+    align-items: baseline;
+    position: relative;
+}
+
+.information__icon {
+    position: absolute;
+    right: -40px;
+    top: 9px;
+    display: block !important;
+    color: gray;
+}
+
+.form__input_select {
+    width: 100% !important;
+}
+
+.icons__container {
+    padding: 5px 16px;
+}
+
+.icons__list {
+    margin: 0;
+    padding: 0;
+    display: grid;
+    grid-template-columns: repeat(auto-fill, minmax(62px, 1fr));
+    list-style: none;
+}
+
+.icons__icon {
+    height: 56px;
+    width: 56px;
+    border-radius: 3px;
+    background-color: transparent;
+}
+
+.icons__icon_btn {
+    padding: 3px;
+    margin: 3px auto;
+    padding-bottom: 0;
+    border:none;
+    background-color: transparent;
+    border: 3px solid transparent;
+    border-radius: 3px;
+    cursor: pointer;
+
+    &:hover {
+        border: 3px solid black;
+    }
+}
+
+.action_button {
+    margin-right: 25px !important;
+}
+
