+++ client/src/redux/rootSaga.ts
@@ -11,7 +11,7 @@ import teamSaga from 'containers/TeamPage/logic/saga';
 import saveFilterSaga from 'containers/SaveFilterModal/logic/saga';
 import usersSaga from 'commonLogic/users/saga';
 import issueSaga from 'pages/IssuePage/logic/saga';
-import projectSaga from 'containers/ProjectSettings/logic/saga';
+import projectSaga from 'containers/ProjectDetails/logic/saga';
 import scrumBoardSaga from 'containers/Board/Scrum/logic/saga';
 import peoplePageSaga from 'containers/People/logic/saga';
 import issueCommentSaga from 'components/IssueCommentForm/logic/saga';
